Add files via upload
The code for the gyro still needs work, but before that happens I have to throw nylon bolts onto the case instead of the metal ones, as they're influencing the magnets inside of the IMU board. They should arrive tomorrow, but it will take some work to figure out appropriate rotation values.

Once the gyro is set up I'll start testing on the RockBlock 9603.